requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
playwright==1.47.0
python-dotenv==1.0.1
openai==1.50.0
httpx==0.27.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.139.0
pydantic==2.7.4